+++ CYHelperDemo/CYHelperDemoTests/CYHelperCollectionUtilsTests.m
@@ -10,11 +10,11 @@
 #import "NSArray+CYHelper.h"
 #import "NSSet+CYHelper.h"
 
-@interface CYHelperMapFilterTests : XCTestCase
+@interface CYHelperCollectionUtilsTests : XCTestCase
 
 @end
 
-@implementation CYHelperMapFilterTests
+@implementation CYHelperCollectionUtilsTests
 
 - (void)setUp
 {
@@ -31,7 +31,7 @@
 - (void)testMapWithBlock {
     NSArray *arr = @[@"hello", @"world", @"this", @"is", @"a", @"test"];
     NSArray *correctArray = @[@5, @5, @4, @2, @1, @4];
-    NSArray *result = [arr mapWithBlock:^id(id value) {
+    NSArray *result = [arr map:^id(id value) {
         return @([value length]);
     }];
     XCTAssertTrue([result isKindOfClass:[NSArray class]], @"array filter with block test faild: result is not a array");
@@ -40,7 +40,7 @@
     }
     
     NSSet *set = [NSSet setWithArray:arr];
-    NSSet *resultSet = [set mapWithBlock:^id(id value) {
+    NSSet *resultSet = [set map:^id(id value) {
         return @([value length]);
     }];
     XCTAssertTrue([resultSet isKindOfClass:[NSSet class]], @"set filter with block test faild: result is not a set");
@@ -53,7 +53,7 @@
 - (void)testFilterWithBlock {
     NSArray *arr = @[@"hello", @"world", @"this", @"is", @"a", @"test"];
     NSArray *correctArray = @[@"hello", @"world", @"this", @"test"];
-    NSArray *result = [arr filterWithBlock:^BOOL(id value) {
+    NSArray *result = [arr filter:^BOOL(id value) {
         return [value length] > 3;
     }];
     XCTAssertTrue([result isKindOfClass:[NSArray class]], @"array filter with block test faild: result is not a array");
@@ -62,7 +62,7 @@
     }
     
     NSSet *set = [NSSet setWithArray:arr];
-    NSSet *resultSet = [set filterWithBlock:^BOOL(id value) {
+    NSSet *resultSet = [set filter:^BOOL(id value) {
         return [value length] > 3;
     }];
     XCTAssertTrue([resultSet isKindOfClass:[NSSet class]], @"set filter with block test faild: result is not a set");
@@ -75,7 +75,7 @@
 - (void)testRejectWithBlock {
     NSArray *arr = @[@"hello", @"world", @"this", @"is", @"a", @"test"];
     NSArray *correctArray = @[@"hello", @"world", @"this", @"test"];
-    NSArray *result = [arr rejectWithBlock:^BOOL(id value) {
+    NSArray *result = [arr reject:^BOOL(id value) {
         return [value length] <= 3;
     }];
     XCTAssertTrue([result isKindOfClass:[NSArray class]], @"array filter with block test faild: result is not a array");
@@ -84,13 +84,23 @@
     }
     
     NSSet *set = [NSSet setWithArray:arr];
-    NSSet *resultSet = [set rejectWithBlock:^BOOL(id value) {
+    NSSet *resultSet = [set reject:^BOOL(id value) {
         return [value length] <= 3;
     }];
     XCTAssertTrue([resultSet isKindOfClass:[NSSet class]], @"set filter with block test faild: result is not a set");
     NSSet *correctSet = [NSSet setWithArray:correctArray];
     for (id value in correctSet) {
         XCTAssertTrue([resultSet containsObject:value], @"set filter with block test faild");
+    }
+}
+
+- (void)testFlattenArray {
+    NSArray *arr = @[@1, @[@2, @3], @[@4, @[@5, @6]]];
+    NSArray *correctArray = @[@1, @2, @3, @4, @5, @6];
+    NSArray *result = [arr flattenArray];
+    XCTAssertTrue([result isKindOfClass:[NSArray class]], @"array filter with block test faild: result is not a array");
+    for (NSInteger i = 0; i < correctArray.count; i++) {
+        XCTAssertEqualObjects(result[i], correctArray[i], @"array filter with block test faild.");
     }
 }